+++ packages/codegen/src/with-cache.ts
@@ -21,6 +21,7 @@ export const withCache =
       circular: LookupEntry,
       ...rest: Other
     ) => T,
+    saveCircularResult: (result: T) => boolean,
   ): FnWithStack<Other, T> =>
   (input, cache, stack, ...rest) => {
     const { id } = input
@@ -28,7 +29,7 @@ export const withCache =
 
     if (stack.has(id)) {
       const res = onEnterCircular(() => cache.get(id)!, input, ...rest)
-      cache.set(id, res)
+      if (saveCircularResult(res)) cache.set(id, res)
       return res
     }
 
@@ -40,6 +41,7 @@ export const withCache =
       result = onExitCircular(result, cache.get(id)!, input, ...rest)
     }
 
-    cache.set(id, result)
+    if (saveCircularResult(result)) cache.set(id, result)
+
     return result
   }